src/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic-settings==2.2.1
aiofiles==23.2.1
langchain==0.1.20
PyMuPDF==1.24.3
motor==3.4.0
pydantic-mongo==2.3.0
openai==1.35.13
cohere==5.5.8
qdrant-client==1.10.1
beautifulsoup4==4.13.4
Requests==2.32.4
# HuggingFace Dependencies - Just sentence-transformers! (LangChain uses it)
sentence-transformers>=2.3.0